Split MyNdlaPage into MyNdlaPage and SupportPage
